requirement.txt
psycopg2-binary==2.9.5
django-cors-headers
